+++ web/callcontract/callcontract-page.component.css.js
@@ -1,5 +1,6 @@
-.outputarea {
+export default /*css*/ `.outputarea {
     background-color: #eef;
     width: 100%;
     padding: 10px;
-}
+}`;
+
