+++ web/callcontract/callcontract-page.component.html.js
@@ -1,13 +1,15 @@
-<h3>Call contract on-chain (JSVM enclave)</h3>
-<p>Here you can call a contract deployed to a JSVM enclave. Standalone contracts can be called like any other NEAR smart contract using e.g. NEAR CLI, or the standalone contract UI in this web app.</p>
+export default /*html*/ `<h3>Call contract on-chain</h3>
+<p>Here you can call any NEAR contract</p>
 <p>
     <mwc-textfield label="Contract name" id="contractnameinput"></mwc-textfield>
     <mwc-textfield label="Method name" id="methodnameinput"></mwc-textfield>
     <mwc-textfield label="arguments (JSON)" id="argsinput"></mwc-textfield>
     <mwc-textfield id="depositinput" label="attached deposit" type="number"></mwc-textfield>
+    <mwc-textfield id="gasinput" label="gas" type="number"></mwc-textfield>
 </p>
 <p>
-    <mwc-button raised id="callcontractbutton">Call contract</mwc-button>
+    <mwc-button raised id="callcontractbutton">Call (with transaction)</mwc-button>
+    <mwc-button raised id="viewcontractmethodbutton">View (no transaction)</mwc-button>
 </p>
 <h3>Contract output</h3>
 <p>
@@ -15,6 +17,6 @@
 <div class="outputarea">
     <pre style="white-space: pre-wrap;"><code id="contractoutput">
 
-                    </code></pre>
+    </code></pre>
 </div>
-</p>
+</p>`;
